Tabular: Removed predictor_legacy.py and tabular_prediction.py
