--- tabular/src/autogluon/tabular/task/tabular_prediction/__init__.py
@@ -1,1 +1,0 @@
-from .tabular_prediction import TabularPrediction
